bsqi uses end time set in HRV params
Previously, when end time was used from the last beat in the annotation. When this was used to get the SQI of a signal with multiple leads by calling bsqi once for each lead, the returned windowed segments variable had a different total length for each lead. Setting the end time to the last timestamp in the ecg will guarantee the same number of windows for each lead.